Revert "refactor: move hooks to separate folder"
This reverts commit 0b25d23e2e58e2ed5203335ba7123c2f7fdfec5d.
